Some fixes to the java classes, and added upVote and downVote methods

+++ Website/src/javaclasses/Ingredient.java
@@ -25,7 +25,7 @@ public class Ingredient {
         this.name = name;
     }
 
-    public float getQuantity() {
+    public double getQuantity() {
         return this.quantity;
     }
